Separate report table from main component.
